+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.1.4
+- [#4](https://github.com/offirgolan/ember-light-table/pull/4) Apply width on cell component based on it's columns width [@steffenbrem](https://github.com/steffenbrem)
+- [#5](https://github.com/offirgolan/ember-light-table/issues/5) Add insertRowAt & insertColumnAt to public Table API
+- Return pushed/inserted rows & columns from public APIs
+
+## 0.1.3
+- Fixed an issue where cell value was not bounc
+
 ## 0.1.2
 - Ability to push rows using ArrayProxy instances